Section 23 - Blog Website
Finished the third boss level challenge. Changed the base style to add a bit more flair. Also made some modifications, like a clickable website logo, a kebab casing for the post names and the fact that an additional "if" statement checks if no matching posts were found, and then redirects the user back to the home route.