Section 32 - Authentication & Security - Lvl 6
Sixth part of the authentication and security section of the course. This section added the ability to use passport to authenticate users using Google and then Facebook (through an additional exercise). Its really simple, but the syntax is a little bit strange. In the get requests for the callback URLs, there isnt a callback function, a call direct call to "passport.authenticate" is first required. Also make sure to correctly setup the developer console for both Google and now "Meta", in order for everything to correctly work.