Section 33 - React - Up to Spread Operator
All exercises from the last commit to the lecture on the spread operator. On the last excercise specifically, Angela uses the spread operator as a way to reduce the code inside the "handleInput()" function. However, I decided to shorten it by using a "deep copy" of the object with all the info. She uses an even more compact method. Committing just to keep track of my previous method.