Section 33 - React - ToDo List / Managing Component Tree
Removed the inputs from the to-do list app, and placed it in a new separate components. Used the ability to pass a function from the App component in order to play inside another component, with variables and states inside the main App component. Also added a small delay between the deletion of a task, for the "crossed out" state of the task to be visible. However, this introduced a bug. After the task is deleted, the next task is crossed out instead (probably because the next element now shares the same ID).